+++ src/index.html
@@ -40,7 +40,7 @@
             </div>
         </div>
     </main>
-    <section class="wrapper__body generations">
+    <section class=" wrapper__body generations">
         <form>
             <label for="pokemons">Choose a Pokemon:</label>
             <select class="generations__list" name="pokemons" id="genga">
@@ -62,14 +62,6 @@
         </section>
     </div>
     <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-    <script type="module" src="./js/index.js"></script>
+    <script src="./js/index.js"></script>
 </body>
 </html>
-
-<div>
-    <h1>
-        hola
-    </h1>
-    <img src="" alt="">
-    <p></p>
-</div>
