--- src/js/models/Pokemons.js
@@ -1,9 +1,0 @@
-class Pokes{
-    constructor(name,id,sprites,types){
-        this.name = name
-        this.id = id
-        this.sprites = sprites
-        this.type = types
-    }
-}
-export default Pokes
